+++ CHANGES.md
@@ -2,8 +2,11 @@
 
 #### next release (8.2.16)
 
+- Make srsName and outputFormat for WFS requests dynamic
 - Added `excludeInactiveDatasets` to `CkanCatalogGroup` (`true` by default). This will filter out CKAN Datasets which have `state` or `data_state` (data.gov.au specific) **not** set to `"active"`.
 - Tweak `generateCatalogIndex` to use less memory. (+ add `diffCatalogIndex.js` script to show added/removed members between two catalog index files)
+- Fix `isTerriaFeatureData` bug - not checking `isJsonObject`
+- Move `RelatedMaps` to terriajs. They are now generated from `configParameters` (see [`doc/customizing/client-side-config.md`](./doc/customizing/client-side-config.md#relatedmap))
 - [The next improvement]
 
 #### 8.2.15 - 2022-09-16